Create auth service New service handles any errors which may occur from the server/rethinkdb. Allows for upper level code to care about result and not errors
