+++ lib/routes/auth/authentication.js
@@ -7,20 +7,17 @@
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  */
-var Authentication = requireFromRoot('server/auth/authentication');
+var Authentication = requireFromRoot('lib/service/auth');
 
 var Routes = {
   post: {
     login: function (request, reply) {
 
-      Authentication.loginUser(request.payload.email, request.payload.password, request.payload.totp_token, request.payload.remoteAddress, function (err, success, data) {
-        if (err === undefined && success) {
+      Authentication.loginUser(request.payload.email, request.payload.password, request.payload.totp_token, request.payload.remoteAddress, function (success, data) {
+        if (success) {
           request.auth.session.set(data);
           reply.redirect('/index');
         } else {
-          if (err !== undefined) {
-            Logger.error(err);
-          }
           request.session.flash('loginError', data, true);
           reply.redirect('/auth/login');
         }
@@ -28,15 +25,7 @@ var Routes = {
 
     },
     totp: function (request, reply) {
-
-      Authentication.isTOTPEnabled(request.payload.check, function (data) {
-        if (data === true) {
-          reply('true');
-        } else {
-          reply('false');
-        }
-      });
-
+      reply(Authentication.isTOTPEnabled(request.payload.check) ? "true" : "false");
     },
     register: function (request, reply) {
